+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.0.1
 ## Changes
 - Add new Intercept Event `Error` triggered on canceling request or exception.
+- Add new types for Intercept Event Callbacks `PostResponseCallback`, `PreRequestCallback`, `PreResponseCallback`, `ErrorCallback`
 
 # 2.0.0
 ## Breaking Changes